fix in CI

+++ tests/test_minimal/test_tools/test_backend_and_dataset_configuration/test_helpers/test_get_default_backend_configuration.py
@@ -130,7 +130,7 @@ epochs/stop_time/data
 
 acquisition/NewRawTimeSeries/data
 ---------------------------------
-  dtype : int32
+  dtype : int64
   full shape of source array : (2, 3)
   full size of source array : 0.00 GB
 
@@ -232,7 +232,7 @@ epochs/stop_time/data
 
 acquisition/NewRawTimeSeries/data
 ---------------------------------
-  dtype : int32
+  dtype : int64
   full shape of source array : (2, 3)
   full size of source array : 0.00 GB
